chore: move mcp-common/src/util/poll.ts to mcp-common/src/poll.ts
This is for consistency with the rest of the helpers in mcp-common
